Creation of Dockerfile and docker-compose
Addition of new help command with pagination and using embeds. This will mean that there's going to be a new package that needs to be added for it to work.
This is also the update that requires at least 3.8.0 to run if you want to use the new help command.