Revamped Weather command
This newer version of the weather command uses openweathermap instead of weatherstack, since weatherstack is more limited when used for free.

This revamped version, together with some other QoL changes (like the addition of a simple emoji that displays the weather description) has the additional feature of showing multiple units when asked.
Simply type ?weather (f|m) (city) and it will display the information in either metric or imperial systems!

New config.py.example model updated with new api

Modified how the data is received to have the timezone applied as well.

This might break some stuff, so please remove ws['timezone'] if the time is off.