Add new command: Reddit.py
This command allows you to cycle through the most popular posts of each subreddit! Just type ?reddit <subreddit> and you'll be filled with nice pictures that made it to the top!

This shuffle option will allow users to search through a list of 20 max pictures that are random and getting popular at that moment.